Updating more of the python scripts
Most of the changes are just putting the #! at the top for run-from-commandline ability.

turn_reminder was updated to use Gmail and is functioning properly now.

orders_status now puts any errors that the script finds into a sp##.ord.log file that can be used to send to the players for order corrections.

I should take some time to go and make the directories and files used be variables that are configured in the top-matter of the file instead of hardcoded inside of the code operations.