Main update is to orders_status.py.  Added email option for error reporting
Added the functionality to email a player if their orders have errors in them.  This will not only generate the log file, but will also send an email with that log file to the player so that they can see where the error is.

usage: python3 orders_status.py -e <anything here triggers it>