Corrected orders_clean.py to stop leaving erroneous data at the end of the file
        try:
            with open(orders, "r+") as f:
                text = f.read()
                f.truncate(0)
                text2 = os.linesep.join([s for s in text.splitlines() if s])
                f.seek(0)
                #f.write(text2)
                if not text2.isspace():
                    f.write(text2)
            f.close()

replaced the original f.write() with the if not text2.isspace() check, then writing that out if it is not a blank line.  Also added the f.truncate(0) after reading in the file to clear the file for writing the cleaned orders back to it.

Also removed some additional print() statements from orders_status.py that was printing out extra debugging data and making the output ugly.